Ignore keys that fail recovery after many attempts.
Some keys in the SKS global pool fail to parse properly with
hockeypuck's openpgp package. These seem to consist primarily of a few
v2 and v3 keys that probably aren't worth the hassle. Hockeypuck will
add the recovered hash to the prefix tree after making N recovery
attempts and failing to discover the digest and load it properly.

This is better than hammering peers for the same unparseable keys
forever. It's also logged, to help debug failure to load a valid, modern
key. Worst case, if we ignore valid keys, we can just recommend a pbuild
from scratch on upgrade.