Match username or domain part of email addresses
Users want to search for keys for a particular username, like “prz”.
Or they want to search for keys for a domain, like “pgp.com”.

Postgres’ `to_tsarray` does not split across the `@`-sign, so email
addresses are indexed as a whole lexeme. This is great to find
“prz@pgp.com” but not good for the above use-cases.

`keywords` now recognizes an email address and splits across the
first `@`-sign it encounters to find the username and domain.

Contributed by Quantcast: https://quantcast.com