Avoid incorrect use of sync.WaitGroup and improve skew
The current code has several issues, including incorrect use of sync.WaitGroup.
What it attempts to achieve is for a readAcquire to start a read/flush
sequence and allow more readers to start reading, taking priority over the
flush (unless full). However this results in various races between p.wg.Add
and p.wg.Wait, which can lead to:

  panic: sync: WaitGroup is reused before previous Wait has returned

This can be triggered by setting p.settings.GossipIntervalSecs to 1,
however it can also occur in normal operation due to the way that the
gossip interval is handled with rand.Intn.

Replace the sync.WaitGroup with a sync.Cond and a separate readers
counter. This allows multiple readers to readAcquire, with the flush being
triggered when the last readRelease drops readers to zero.

Also avoid modifying p.once while only holding the p.mu as a read lock.

Rework the gossip interval skew so that it is +/-10% of the configured
value, rather than allowing it to potentially occur zero seconds after
the last gossip. This avoids constant gossip while still achieving a
spread of the interval.