Handle keywords that exceed tsvector limits.
Currently there are a number of bogus keys that result in keywords being
generated that exceed PostgreSQL's tsvector limits. This means that the
key then fails to insert. Detect this condition and simply omit the
keywords for this key - this allows us to insert the key and simply
means that it is only findable by fingerprint.