Relocate go.mod, vendored dependencies, update goleveldb.
This started with upgrading goleveldb to see if memory issues
there might be addressed in a later release. Looking at closed issues
there, it seems that some memory issues have been addressed there.

Discovered that the project no longer builds with go 1.13 and newer.
Needed to move go.mod into the top-level package directory (was in
$GOPATH/src). That also requires the vendor directory to move with it.