docker-compose: Allow using a non-default CA
The SSL certificate for HTTPS is managed by certbot. The default
CA is Let's Encrypt, but some environments may not use it
(e.g. because they are not reachable from the public internet).

Thankfully, certbot implements a --server option to override the
default. Add support for this option to init-letsencrypt.bash and
add the ACME_SERVER variable to settings.